Extract calculate highlights into separate file
